Extended package form
Extend (package <name>) to support import, import-as and export
clauses. There is an example in small-coalton-programs --
fact-fib.coal -- and documentation in doc/package.md.

This syntactic extension supports the development of .coal files,
extended emacs support, and exploration of the semantics of package
forms, but is not yet a public API feature.