Remove environment-map
Environment maps were introduced in order to provide platform-agnostic
dumping of hashtables present in Coalton environment structures.

A review of the three locations where these maps are
present (method-codegen-sym, class-variable-map, superclass-map)
indicates that in all cases, a simple list or alist structure is more
appropriate because of the small number of elements consulted.

In particular, method-codegen-syms are 1-1 with the method-names list,
so no mapping is needed at all.